Updated SubCommand to now create player-specific and faction-specific commands.
Also removed "isPlayer(CommandSender)" as its been replaced using the player-specific command functionality.

Added EphermeralData instance to the subcommand class, under "ephemeral" so it can be contacted when required.

Added a utility method called "checkPermissions(Sender, Permissions)", which will check for an array, ensuring that the sender has permission, if they do not, an error message is sent.