Fixed #1114.
Java16 brought a lot of changes, some of which have been about since 1.8 and even before that too! I have therefore transitioned the code from assuming we have an array of String, to forcefully creating one. This (despite the sound of it) changes no functionality, apart from fixing the bug and allowing this to work with Java16 on 1.17.X and Java1.8 on 1.8.9 (Both tested to be working on my end).