Contiguous claims config option implemented
Default is set to false, to keep the on-install functionality the same as it is now. I didn't really know whether or not to put the function that checks adjacency in the MfClaimService class or the MfFactionClaimCommand class, but it works nonetheless. Will gladly change if the reviewer would rather see it changed.

I know it's a bit messy and hard to read, please let me know how to improve the readability!