Expanded Territory Title Notifications (#1736)
This PR expands upon the plugins call to the player.sendTitle method for
territory notifcations.

- Territory indicator now includes the factions description, displayed
as a subtitle. This brings it to parity with many other factions
plugins. By default, it is also configured to be much shorter, as the
title can become distracting in combat scenarios.

- The length of the title's fadeIn, duration, and fadeOut can now be
controlled with a config.yml option.

Note that this creates an issue as there is no explicit length limit to
the factions description, which may cause certain factions descriptions
to go offscreen when being displayed as a title. Perhaps this might be
limited, or another string could be stored in a faction, such as a
tagline?