Config Option - Only Render Territories Upon Startup (#1768)
## Problem
Rendering territories is causing a lot of lag for some servers.

## Solution
This PR introduces a config option to prevent the rendering of
territories after the initial startup of the server. This limits the lag
spike to one instance per session.

## Testing
This was used by a server owner in the community and their TPS improved
from 15 to 18. This seems to ease the lag a bit.

## Related Issues
closes #1767