Added Brazilian Portuguese (pt-BR) translation for Faction-Medieval p… (#1758)
I've added a Brazilian Portuguese (pt-BR) translation for the
Faction-Medieval plugin. The translation is contained in a new file
within the language folder to support Portuguese-speaking users.

I believe this translation will help expand the user base of the plugin
to those who prefer or need to play in Portuguese. Please review the
changes, and let me know if there are any modifications or improvements
needed.

Thank you for considering this contribution!

Best regards,
MestreWilll

---------

Co-authored-by: Daniel McCoy Stephenson <dmccoystephenson@gmail.com>
Co-authored-by: Ren Binden <ren.binden@gmail.com>
Co-authored-by: Donut <86391164+Chocolate1Donut@users.noreply.github.com>
Co-authored-by: MestreWilll <WillXOkumura@hotmail.com>