Test Server Dockerfile (#1775)
## Docker
A Dockerfile has been added for a test server image. This can be spun up
using the compose.yml file with `docker-compose up`

## Aliases in plugin.yml
Some aliases in the plugin.yml were modified to not use special
characters, as this can result in the plugin being unusable on some
servers due to the plugin.yml being unable to be parsed.