Enable Bluemap in Test Server (#1828)
## Problem  
The test server must be able to run Bluemap before Bluemap support can
be implemented.

## Solution  
A new environment variable has been added to `sample.env` to enable the
Bluemap plugin. When set to `true`, Bluemap will be installed in the
test container.

## Testing  
After starting the test container, Bluemap was verified to be accessible
at `localhost:8100`.